+++ index.js
@@ -42,3 +42,4 @@ exports.isBoolean = function(value) { return typeof value == "boolean" }
 exports.isNumber = function(value) { return typeof value == "number" }
 exports.isString = function(value) { return typeof value == "string" }
 exports.isFunction = function(value) { return typeof value == "function" }
+exports.isArray = Array.isArray